requirements.txt
nltk==3.9.1
PyMuPDF==1.26.4
python-docx==1.1.2
pytesseract==0.3.13
pdf2image==1.17.0
jellyfish==1.1.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
